refactor(lightgbm): extract socket lifecycle support
## Summary
Move port reservation and cleanup implementations into a package-private support object while retaining the existing NetworkManager entry points. This keeps the integrated networking implementation below the repository's 800-line source-file limit without changing behavior.

## Prompting Intent
Rebase PR #2612 after #2595 merged, preserve both LightGBM networking fixes, resolve resulting CI failures, and keep the implementation lean, reviewable, and fully validated.

## Linked Sources
- Rebased retry-topology PR: https://github.com/microsoft/SynapseML/pull/2612
- Merged port-reservation PR: https://github.com/microsoft/SynapseML/pull/2595
- Failed style check: https://github.com/microsoft/SynapseML/actions/runs/31225571403/job/93019061030

## Rationale
The two fixes are behaviorally complementary but together made NetworkManager.scala 909 lines. Extracting cohesive socket lifecycle mechanics avoids a scalastyle suppression or behavior reduction, preserves package-level test seams through thin delegates, and leaves NetworkManager.scala below 800 lines. Combined style checks and all 31 networking tests pass.

Copilot-Session: 81d39bfc-927c-418a-90a8-e0f2cd8fc128
